src/requirements.txt
arrow==1.1.1
asgiref==3.4.1
aws-psycopg2==1.2.1
backports.zoneinfo==0.2.1
botocore==1.21.65
cement==2.8.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.4
colorama==0.4.3
dataclasses==0.6
dateparser==1.0.0
Django==3.2.7
django-picklefield==3.0.1
djangorestframework==3.12.4
freezegun==1.1.0
future==0.16.0
idna==3.2
importlib-metadata==4.11.2
install==1.3.4
jmespath==0.10.0
joblib==1.1.0
mock==4.0.3
moment==0.12.1
Moments==2.0
nltk==3.7
oauthlib==3.1.1
pathspec==0.5.9
psycopg==3.0.7
psycopg2==2.9.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
requests-oauthlib==1.3.0
schedule==1.1.0
semantic-version==2.8.5
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
textblob==0.17.1
times==0.7
tqdm==4.63.0
tweepy==3.10.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
wcwidth==0.1.9
zipp==3.7.0
